Add files via upload
You need to come up with a story in Tableau that talks about the level of demand in each
centre. This analysis needs to be granular enough to include product information as well. The
client wants an end-to-end report to understand which fulfilment areas are doing well and
which aren't. You can also talk about centre-meal combinations to add nuance to your final
submission.